+++ speechain/datasets/pyscripts/__init__.py
@@ -1,0 +1,9 @@
+"""Executable scripts shared by all dataset-dumping pipelines.
+
+Each script in this sub-package is designed to be executed directly, e.g.::
+
+    python ${SPEECHAIN_ROOT}/speechain/datasets/pyscripts/feat_extractor.py ...
+
+See ``${SPEECHAIN_ROOT}/datasets/data_dumping.sh`` for how the scripts are
+chained into a complete dataset-dumping pipeline.
+"""